local docker caching example
